+++ src/Trulioo.Client.V1/Model/RecordRule.cs
@@ -2,7 +2,21 @@
 {
     public class RecordRule
     {
+        /// <summary>
+        /// Name of RecordRule.
+        /// </summary>
         public string RuleName { get; set; }
+        /// <summary>
+        /// Rule Description.
+        /// </summary>
         public string Note { get; set; }
+        /// <summary>
+        /// Name of the Secondary RecordRule.
+        /// </summary>
+        public string SecondaryRuleName { get; set; }
+        /// <summary>
+        /// Secondary Rule Description.
+        /// </summary>
+        public string SecondaryRuleNote { get; set; }
     }
 }