+++ src/Trulioo.Client.V1/Model/TestEntityDataFields.cs
@@ -1,0 +1,10 @@
+﻿namespace Trulioo.Client.V1.Model
+{
+    public class TestEntityDataFields : DataFields
+    {
+        /// <summary>
+        /// Friendly name for the test entity
+        /// </summary>
+        public string TestEntityName { get; set; }
+    }
+}